requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
langchain-experimental>=0.0.40
langchain-text-splitters>=0.0.1
langgraph>=0.0.20
langsmith>=0.0.65
chromadb>=0.4.22
sentence-transformers>=2.2.2
pypdf>=3.17.1
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=10.1.0
duckduckgo-search>=3.9.5
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.1
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
deepeval>=0.20.0
fastapi>=0.104.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
tqdm>=4.66.1
ollama>=0.1.0